api/requirements.txt
# Azure Functions for Python
azure-functions>=1.11.0

# Face recognition library
face-recognition>=1.3.0

# Dependencies for face-recognition
dlib>=19.24.0
numpy>=1.21.0
Pillow>=9.0.0

# Azure Storage SDK
azure-storage-blob>=12.14.0

# Database driver
pyodbc>=4.0.35
